main: bring back default loop and stat/mon
This switches main.c / main thread back to the form it had a few
years ago, where it had a libev loop and that loop/thread also
runs all the monitoring and stats code.  This basically undoes the
core work of abb5c635 , but updated a bit for the changes since
that time, and doesn't undo the removal of the libev io and
timeout collection intervals.

This will allow us to set up a control socket in the main thread
as well, which can easily get access to stats...