Stats: only JSON, and only via control socket
This removes from the stats code (and support elsewhere) for:
  * CSV output format
  * HTML output format
  * HTTP listener
  * Syslog emitter

The JSON output format is retained, but is only accessible via the
control socket.  "gdnsdctl stats" dumps the current stats in JSON
form from the control socket to stdout.  Periodic file output can
be driven by a cronjob executing gdnsdctl if needs be, or another
tool that attaches directly to the control socket.

With the stats HTTP listener gone, a few remaining optimizations
for later and simpler DNS socket binding are achieved along the
way as well.