Remove DYNx TTL-mangling in core code
This gets rid of the the monitor- and/or state- based TTL
reduction and clamping code in the runtime data structures and
code, as well as the zonefile scanners.  Note the rest of the
infrastructure for it is still there in plugin and monitoring
code, but it has no visible effect on DNS responses with this
commit.

This breaks various tests that make use of the TTL-range stuff,
but I'll wait to fix them until we clean up everything
plugin-related...